pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = ["src/**/*.py"]

[tool.isort]
profile = "black"
line_length = 120
skip_glob = ["submodules/*"]

[tool.black]
line-length = 120
force-exclude = "submodules/*"

[project]
name = "HyperGaussians-FlashAvatar"
version = "1.0.0"
requires-python = ">=3.9"
dependencies = [
  "av==12.0.0",
  "black==24.10.0",
  "chumpy==0.70",
  "dearpygui==2.0.0",
  "einops",
  "isort==5.13.2",
  "loguru==0.7.3",
  "lpips==0.1.4",
  "matplotlib==3.10.0",
  "numpy==1.23.5",
  "opencv-python==4.11.0.86",
  "plotly==5.24.1",
  "plyfile==1.1",
  "pre-commit==4.0.1",
  "pytorch3d@git+https://github.com/facebookresearch/pytorch3d.git@V0.7.8",
  "roma==1.5.1",
  "scikit-image==0.24.0",
  "scipy==1.13.1",
  "tabulate==0.9.0",
  "timm==1.0.14",
  # "torch", # Manually install to match CUDA Version
  "torchmetrics==1.6.0",
  # "torchvision", # Manually install to match CUDA Version
  "tyro==0.9.1",
  "wandb==0.18.7",
  "yacs==0.1.8"
]
authors = [
  {name = "Gent Serifi"},
  {name = "Marcel C. Buehler"}
]

description = "HyperGaussians: High-Dimensional Gaussian Splatting for High-Fidelity Animatable Face Avatars"
readme = "README.md"
